+++ src/atlas_sprites/render_context/tests.rs
@@ -291,7 +291,7 @@ fn update_sprite_values() {
         // update_sprite_values
         assert_ne!(
             texture_atlas.index,
-            render_context.image_font.atlas_character_map[&first_char].atlas_index
+            render_context.image_font.atlas_character_map[&first_char].character_index
         );
         assert_ne!(color, render_context.image_font_sprite_text.color);
 
@@ -300,7 +300,7 @@ fn update_sprite_values() {
         // Verify the texture atlas and color are updated
         assert_eq!(
             texture_atlas.index,
-            render_context.image_font.atlas_character_map[&first_char].atlas_index
+            render_context.image_font.atlas_character_map[&first_char].character_index
         );
         assert_eq!(color, render_context.image_font_sprite_text.color);
     });