+++ CHANGELOG.md
@@ -12,14 +12,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `LetterSpacing` enum to allow customizable kerning between characters.
 - `letter_spacing` field to `ImageFontSpriteText` for individual entity customization.
 - Example in `atlased_sprite` to demonstrate `LetterSpacing` usage.
-- Basic support for loading fonts from `.fnt` files.
-  - Introduced `ImageFntFontLoader` for parsing `.fnt` files into `ImageFont` assets.
-  - Added a new `fnt` feature flag to enable `.fnt` font loading.
+- Support for fonts that span multiple texture pages.
+- `BmFontLoader` for loading bitmap font file formats (`txt.fnt`, `xml.fnt`, `bin.fnt`).
 
 ### Changed
 
 - Updated text rendering calculations to account for `letter_spacing`.
 - No longer repeatedly prints error for missing font assets.
+- `ImageFont` now supports multiple textures instead of a single one. (Currently only supported by `.fnt` definitions)
+
+### Removed
+
+- `ImageFntFontLoader` font loader, replaced by `BmFontLoader`.
 
 ### Breaking Changes